fix route.php

+++ app/Controllers/dashboard/AdminOlim.php
@@ -91,18 +91,14 @@ class AdminOlim extends BaseController
             if ($this->is_pass_same($password)) {
                 return redirect()->to('dashboard/admin-olim/verify-konfirmasi-team/' . $id . '/1');
             } else {
-                $data_status = [
-                    'olim_tim_id' => $team['olim_tim_id'],
-                    'olim_status' => 1
-                ];
-                $this->Olim_Model->save($data_status);
+
                 $data = [
                     'user_lomba'       => 'olim',
                     'user_tim_nama'  => $team['olim_tim_nama'],
                     'user_username'    => 'olimpiade' . '_' . $team['olim_tim_nama'],
                     'user_password'    => password_hash($password, PASSWORD_DEFAULT)
                 ];
-                $this->User_Model->save($data);
+
                 $subject = "[Accepted] Olimpiade";
                 $message = "Halo {$team['olim_tim_nama']} dari {$team['olim_asal_institusi']},<br>
                   <br>
@@ -117,8 +113,25 @@ class AdminOlim extends BaseController
                   Salam Hormat,<br>
                   <br>
                   A Renewal Agents 4.0";
-                $this->sendEmail($team['olim_email_ketua'], $subject, $message);
-                $this->session->setFlashdata('msg', 'Berhasil Menerima Tim: ' . $team['olim_tim_nama']);
+                // $this->sendEmail($team['olim_email_ketua'], $subject, $message);
+                $email = \Config\Services::email();
+                $email->setTo($team['olim_email_ketua']);
+                $email->setFrom('arenewalagent@gmail.com', 'ARA 4.0');
+                $email->setSubject($subject);
+                $email->setMessage($message);
+                if ($email->send()) {
+                    $data_status = [
+                        'olim_tim_id' => $team['olim_tim_id'],
+                        'olim_status' => 1
+                    ];
+                    $this->Olim_Model->save($data_status);
+                    $this->User_Model->save($data);
+                    $this->session->setFlashdata('msg', 'Berhasil Menerima Tim: ' . $team['olim_tim_nama']);
+                } else {
+                    // $dd = $email->printDebugger(['headers']);
+                    // $this->session->setFlashdata('msg', $dd);
+                    $this->session->setFlashdata('msg', 'Server Error');
+                }
             }
         } else {
             $kp_path = 'uploads/olim/suket/';
@@ -126,32 +139,40 @@ class AdminOlim extends BaseController
             $tt_path = 'uploads/olim/tiktok/';
             $bukti_bayar_path = 'uploads/olim/bukti_bayar/';
 
-            $this->delete_file($kp_path, $team['olim_kp_surket_ketua']);
-            $this->delete_file($kp_path, $team['olim_kp_surket_anggota_1']);
-
-            $this->delete_file($ig_path, $team['olim_ig_ara_ketua']);
-            $this->delete_file($ig_path, $team['olim_ig_ara_anggota_1']);
-
-            $this->delete_file($tt_path, $team['olim_tiktok_ketua']);
-            $this->delete_file($tt_path, $team['olim_tiktok_anggota_1']);
-
-            $this->delete_file($bukti_bayar_path, $team['olim_bukti_bayar']);
-
-            $this->Olim_Model->where('olim_tim_id', $id)->delete();
             $subject = "[Rejected] Olimpiade";
             $message = "Halo {$team['olim_tim_nama']} dari {$team['olim_asal_institusi']},<br>
-      <br>
-      Terima kasih sudah mendaftar pada event kami, \"Olimpiade.\"<br>
-      <br>
-      Mohon maaf, persyaratan yang anda kirimkan tidak cukup. Mohon mendaftar kembali di halaman registrasi.<br>
-      <br>
-      <br>
-      --<br>
-      Salam Hormat,<br>
-      <br>
-      A Renewal Agents 4.0";
-            $this->sendEmail($team['olim_email_ketua'], $subject, $message);
-            $this->session->setFlashdata('msg', 'Berhasil Menolak Tim: ' . $team['olim_tim_nama']);
+            <br>
+            Terima kasih sudah mendaftar pada event kami, \"Olimpiade.\"<br>
+            <br>
+            Mohon maaf, persyaratan yang anda kirimkan tidak cukup. Mohon mendaftar kembali di halaman registrasi.<br>
+            <br>
+            <br>
+            --<br>
+            Salam Hormat,<br>
+            <br>
+            A Renewal Agents 4.0";
+            $email = \Config\Services::email();
+            $email->setTo($team['olim_email_ketua']);
+            $email->setFrom('arenewalagent@gmail.com', 'ARA 4.0');
+            $email->setSubject($subject);
+            $email->setMessage($message);
+            if ($email->send()) {
+                $this->delete_file($kp_path, $team['olim_kp_surket_ketua']);
+                $this->delete_file($kp_path, $team['olim_kp_surket_anggota_1']);
+
+                $this->delete_file($ig_path, $team['olim_ig_ara_ketua']);
+                $this->delete_file($ig_path, $team['olim_ig_ara_anggota_1']);
+
+                $this->delete_file($tt_path, $team['olim_tiktok_ketua']);
+                $this->delete_file($tt_path, $team['olim_tiktok_anggota_1']);
+
+                $this->delete_file($bukti_bayar_path, $team['olim_bukti_bayar']);
+
+                $this->Olim_Model->where('olim_tim_id', $id)->delete();
+                $this->session->setFlashdata('msg', 'Berhasil Menolak Tim: ' . $team['olim_tim_nama']);
+            } else {
+                $this->session->setFlashdata('msg', 'Server Error');
+            }
         }
         return redirect()->to('dashboard/admin-olim/konfirmasi-team');
     }
